+++ test/data.jl
@@ -10,3 +10,9 @@ end
 
     @test pm.data == math
 end
+
+@testset "check_switch_state_feasibility" begin
+    mn_eng = make_multinetwork(parse_file("../test/data/ieee13_feeder.dss"))
+
+    @test all(check_switch_state_feasibility(mn_eng))
+end